requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python
mediapipe
librosa
soundfile
torch
torchvision
torchaudio
pydantic
python-dotenv
